Introduce a QueuedThreadPool to try to utilise multiple cores
I was running https://github.com/garethr/jruby-embedded-jetty and noted that only one core seems to get used. I'm suggesting this change in the hope that it will use all of the available cores where possible.

It might be that the application is purely IO-bound and that's why I only see a single core running hot, but I don't know jetty well enough to say for sure.